--- web/src/components/Map/ControlPanel/index.scss
@@ -1,15 +1,0 @@
-.filters-container {
-  .ui.icon.input {
-    width: 250px;
-  }
-
-  .mapboxgl-ctrl-group {
-    background: initial;
-    &:not(:empty) {
-      box-shadow: none;
-    }
-    & > button {
-      background-color: #fff;
-    }
-  }
-}
