+++ web/src/components/Map/ControlPanel/types.ts
